Move SplinePathSegment into SplinePathBuilder
